SVG: stop exposing modules on other classes (#10057)
